Validations passing

+++ tests/test_spline_profile.py
@@ -6,7 +6,7 @@ from src.python.field_profile_mlp import SplineProfile
 def test_spline_profile():
     print("--- Testing SplineProfile ---")
     # Use a high-resolution grid near zero to check the derivative
-    rho = torch.linspace(0.0, 10.0, 2000)
+    rho = torch.linspace(0.0, 10.0, 2000, dtype=torch.float64)
     target_flux = 2.0 * np.pi * 0.4
     
     # Initialize Spline Profile
@@ -18,7 +18,7 @@ def test_spline_profile():
     calculated_flux = 2.0 * np.pi * torch.sum(B.squeeze() * rho * dr)
     print(f"Target Flux:     {target_flux:.6f}")
     print(f"Calculated Flux: {calculated_flux.item():.6f}")
-    assert torch.allclose(torch.tensor(target_flux), calculated_flux, rtol=1e-2), "Flux not conserved!"
+    assert torch.allclose(torch.tensor(target_flux, dtype=torch.float64), calculated_flux, rtol=1e-2), "Flux not conserved!"
 
     # 2. Verify Continuity (Visual/Numerical)
     # Check for NaNs or Inf
